Switch more of VCC to RFC8941 number parsing

+++ lib/libvcc/vcc_compile.h
@@ -433,7 +433,7 @@ void Resolve_Sockaddr(struct vcc *tl, const char *host, const char *defport,
 double vcc_DurationUnit(struct vcc *);
 void vcc_ByteVal(struct vcc *, VCL_INT *);
 void vcc_Duration(struct vcc *tl, double *);
-unsigned vcc_UintVal(struct vcc *tl);
+uint64_t vcc_UintVal(struct vcc *tl);
 int vcc_IsFlag(struct vcc *tl);
 int vcc_IsFlagRaw(struct vcc *, const struct token *, const struct token *);
 char *vcc_Dup_be(const char *b, const char *e);
